Moved ExpressionWidget header to a separate component
