+++ frontend/src/metabase/query_builder/components/AggregationPopover/AggregationPopover.styled.tsx
@@ -3,10 +3,6 @@ import AccordionList from "metabase/core/components/AccordionList";
 import { color } from "metabase/lib/colors";
 import FieldList from "../FieldList";
 
-export const ExpressionPopoverRoot = styled.div`
-  min-width: 350px;
-`;
-
 export const AggregationItemList = styled(AccordionList)`
   color: ${color("summarize")};
 `;
