+++ frontend/src/metabase/query_builder/components/expressions/ExpressionWidget.styled.tsx
@@ -68,16 +68,3 @@ export const ActionButtonsWrapper = styled.div`
   display: flex;
   gap: 1rem;
 `;
-
-export const Header = styled.div`
-  color: ${color("text-medium")};
-  padding: 1rem 1.5rem;
-  border-bottom: 1px solid ${color("border")};
-
-  display: flex;
-  align-items: center;
-`;
-
-export const HeaderButton = styled(Button)`
-  color: ${color("text-light")};
-`;
